9. When you roll the dice and make a selection, you are not able to keep the score you selected and move onto the next round.
Create an updateScore function to add this functionality. Your function will need two parameters for the user selected score option. The first parameter will be passed the value of the radio button, remember this is a string, and the second parameter will be passed the id value of the radio button, which is the type of score they achieved.

The function will need to add the user selected value to the score, update the total score text on the page, and add a new li element to the score history ul element, using the format ${achieved} : ${selectedValue} for the li element content.