13. If the user rolls three of one number, and two of another number, this is called a full house. Declare a detectFullHouse function that accepts a single argument. The function will be passed the diceValuesArr array when called.
Your detectFullHouse function should check if the user has rolled three of one number and two of another number. If so, it should update the third radio button to display a score of 25, with the correct attributes. Regardless, it should always update the last radio button to display a score of 0, with the correct attributes.

Don't forget to call your new function when the dice are rolled.